Split off load-pyodide.js from pyodide.js
