+++ src/js/module.js
@@ -1,0 +1,5 @@
+export let Module = {};
+Module.noImageDecoding = true;
+Module.noAudioDecoding = true;
+Module.noWasmDecoding = false; // we preload wasm using the built in plugin now
+Module.preloadedWasm = {};
